db change

+++ fb/models.py
@@ -11,9 +11,9 @@ class Campaign(models.Model):
 	name = models.CharField(max_length=200, default='')
 	campaign_id = models.CharField(max_length=200, default='')
 	status = models.CharField(max_length=200, default='')
-	clicks = models.IntegerField(default=0)
+	clicks = models.BigIntegerField(default=0)
 	cpc = models.FloatField(default=0)
-	impressions = models.IntegerField(default=0)
+	impressions = models.BigIntegerField(default=0)
 	account = models.ForeignKey(Account)
 	
 	def __str__(self):
